requirements.txt
Django==1.10.6
django-appconf==1.0.2
django-imagekit==4.0
olefile==0.44
pilkit==2.0
Pillow==4.0.0
django-countries==4.1
django-mptt==0.8.7
django-smtp-ssl==1.0
djangorestframework==3.6.2